requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.16
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
oauthlib==3.2.2
packaging==24.1
pillow==11.0.0
pip==24.2
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
pytest==8.3.3
python-dotenv==1.0.1
redis==5.1.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
